Replace all the contract calls
